Extract OutputManager
