+++ src/server/frontend/wayland/CMakeLists.txt
@@ -11,6 +11,7 @@ set(
   wayland_executor.cpp          wayland_executor.h
   null_event_sink.cpp           null_event_sink.h
   basic_surface_event_sink.cpp  basic_surface_event_sink.h
+  output_manager.cpp            output_manager.h
 )
 
 add_library(
